--- src/main/java/com/memo/common/JwtProperties.java
@@ -1,15 +1,0 @@
-package com.memo.common;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Component
-@ConfigurationProperties(prefix = "jwt")
-@Getter
-@Setter
-public class JwtProperties {
-	private String secretKey;
-}
