--- src/main/java/com/memo/common/Role.java
@@ -1,7 +1,0 @@
-package com.memo.common;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum Role {
-	GUEST, USER, ADMIN
-}
