+++ src/main/java/com/memo/common/properties/JwtProperties.java
@@ -1,0 +1,15 @@
+package com.memo.common.properties;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Component
+@ConfigurationProperties(prefix = "jwt")
+@Getter
+@Setter
+public class JwtProperties {
+	private String secretKey;
+}
