--- src/main/java/com/memo/login/oauth/OAuthService.java
@@ -1,7 +1,0 @@
-package com.memo.login.oauth;
-
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-public interface OAuthService {
-	OAuthUser service(OAuth2User oAuth2User);
-}
