--- src/main/java/com/memo/login/oauth/OAuthUser.java
@@ -1,8 +1,0 @@
-package com.memo.login.oauth;
-
-public interface OAuthUser {
-	String getEmail();
-	String getUsername();
-	String getId();
-	String getProfileImg();
-}
