--- src/main/java/com/memo/login/oauth/kakao/KakaoTokenResponse.java
@@ -1,21 +1,0 @@
-package com.memo.login.oauth.kakao;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class KakaoTokenResponse {
-	@JsonProperty("token_type")
-	private String tokenType;
-	@JsonProperty("access_token")
-	private String accessToken;
-	@JsonProperty("expires_in")
-	private Integer expiresIn;
-	@JsonProperty("refresh_token")
-	private String refreshToken;
-	@JsonProperty("refresh_token_expires_in")
-	private Integer refreshTokenExpiresIn;
-}
