+++ src/main/java/com/memo/user/entity/Role.java
@@ -1,0 +1,5 @@
+package com.memo.user.entity;
+
+public enum Role {
+	GUEST, USER, ADMIN
+}
