+++ src/main/java/com/memo/user/oauth/OAuthService.java
@@ -1,0 +1,7 @@
+package com.memo.user.oauth;
+
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+public interface OAuthService {
+	OAuthUser service(OAuth2User oAuth2User);
+}
