+++ src/main/java/com/memo/user/oauth/OAuthUser.java
@@ -1,0 +1,8 @@
+package com.memo.user.oauth;
+
+public interface OAuthUser {
+	String getEmail();
+	String getUsername();
+	String getId();
+	String getProfileImg();
+}
