+++ src/main/java/com/memo/user/oauth/kakao/KakaoTokenInfo.java
@@ -1,0 +1,14 @@
+package com.memo.user.oauth.kakao;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Getter;
+
+@Getter
+public class KakaoTokenInfo {
+	private Long id;
+	@JsonProperty("expires_in")
+	private Integer expiresIn;
+	@JsonProperty("app_id")
+	private Integer appId;
+}
