+++ src/main/java/com/memo/user/oauth/kakao/KakaoUserInfo.java
@@ -1,0 +1,37 @@
+package com.memo.user.oauth.kakao;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Getter;
+
+@Getter
+public class KakaoUserInfo {
+	@JsonProperty("kakao_account")
+	private KakaoAccount kakaoAccount;
+	private Long id;
+
+	@Getter
+	static class KakaoAccount {
+		private KakaoProfile profile;
+		private String email;
+	}
+
+	@Getter
+	static class KakaoProfile {
+		private String nickname;
+	}
+
+
+	public String getEmail() {
+		return kakaoAccount.getEmail();
+	}
+
+
+	public String getNickName() {
+		return kakaoAccount.getProfile().getNickname();
+	}
+
+	public String getId() {
+		return String.valueOf(id);
+	}
+}
